+++ qom/utils/misc.py
@@ -162,11 +162,11 @@ def get_limits(mini, maxi, res=2):
         _maxi *= - 1
 
     # update multiplier
-    if _mini != 0 and np.log10(_mini) < res:
+    if _mini != 0 :
         _mult_min = 10**(np.ceil(-np.log10(_mini)) + res - 1)
-    if _maxi != 0 and np.log10(_maxi) < res:
+    if _maxi != 0:
         _mult_max = 10**(np.ceil(-np.log10(_maxi)) + res - 1)
-    _mult = max(10**res, min(_mult_min, _mult_max))
+    _mult = min(10**res, min(_mult_min, _mult_max))
 
     # round off
     _mini = np.floor(mini * _mult) / _mult